Update model parameters and fix embedding method bug

+++ app/functions.py
@@ -460,15 +460,12 @@ def select_embedding_method(embedding_method: str, Channels_next: int, data_to_e
         embedded_blocks = tsne.fit_transform(data_to_embed)
 
     elif embedding_method == "LLE":
-        lle = LocallyLinearEmbedding(
-            n_components=Channels_next,
-            n_neighbors=int(data_to_embed.shape[0] / Channels_next),
-        )
+        lle = LocallyLinearEmbedding(n_components=Channels_next,n_neighbors=int(data_to_embed.shape[0] / Channels_next))
         embedded_blocks = lle.fit_transform(data_to_embed)
     else:
         print("Error: No embedding selected.")
 
-    return embedded_blocks.astype(np.float32)
+    return embedded_blocks.astype(np.float16)
 
 
 def calculate_average_accuracy_kernelCNN(main_function, arguments, dataset_name, iterations=10):
